Extract client from main
Adds on_request and on_notification hooks to avoid some dependencies
