+++ core/main.py
@@ -1,8 +1,6 @@
 import html
-import json
 import os
 import subprocess
-import threading
 import webbrowser
 from collections import OrderedDict
 
@@ -25,6 +23,7 @@ from .configuration import (
     ClientConfig, settings, client_configs, load_settings, unload_settings, PLUGIN_NAME
 )
 from .logging import debug, exception_log, server_log
+from .client import Client
 
 
 SUBLIME_WORD_MASK = 515
@@ -64,174 +63,6 @@ symbol_kind_names = {
 
 
 completion_item_kind_names = {v: k for k, v in CompletionItemKind.__dict__.items()}
-
-
-def format_request(payload: 'Dict[str, Any]'):
-    """Converts the request into json and adds the Content-Length header"""
-    content = json.dumps(payload, sort_keys=False)
-    content_length = len(content)
-    result = "Content-Length: {}\r\n\r\n{}".format(content_length, content)
-    return result
-
-
-class Client(object):
-    def __init__(self, process, project_path):
-        self.process = process
-        self.stdout_thread = threading.Thread(target=self.read_stdout)
-        self.stdout_thread.start()
-        self.stderr_thread = threading.Thread(target=self.read_stderr)
-        self.stderr_thread.start()
-        self.project_path = project_path
-        self.request_id = 0
-        self.handlers = {}  # type: Dict[int, Callable]
-        self.capabilities = {}  # type: Dict[str, Any]
-
-    def set_capabilities(self, capabilities):
-        self.capabilities = capabilities
-
-    def get_project_path(self):
-        return self.project_path
-
-    def has_capability(self, capability):
-        return capability in self.capabilities
-
-    def get_capability(self, capability):
-        return self.capabilities.get(capability)
-
-    def send_request(self, request: Request, handler: 'Callable'):
-        self.request_id += 1
-        debug('request {}: {} '.format(self.request_id, request.method))
-        if handler is not None:
-            self.handlers[self.request_id] = handler
-        self.send_payload(request.to_payload(self.request_id))
-
-    def send_notification(self, notification: Notification):
-        debug('notify: ' + notification.method)
-        self.send_payload(notification.to_payload())
-
-    def kill(self):
-        self.process.kill()
-
-    def send_payload(self, payload):
-        try:
-            message = format_request(payload)
-            self.process.stdin.write(bytes(message, 'UTF-8'))
-            self.process.stdin.flush()
-        except BrokenPipeError as err:
-            sublime.status_message("Failure sending LSP server message, exiting")
-            exception_log("Failure writing payload", err)
-            self.process.terminate()
-            self.process = None
-
-    def read_stdout(self):
-        """
-        Reads JSON responses from process and dispatch them to response_handler
-        """
-        ContentLengthHeader = b"Content-Length: "
-
-        running = True
-        while running:
-            running = self.process.poll() is None
-
-            try:
-                content_length = 0
-                while True:
-                    header = self.process.stdout.readline()
-                    if header:
-                        header = header.strip()
-                    if not header:
-                        break
-                    if header.startswith(ContentLengthHeader):
-                        content_length = int(header[len(ContentLengthHeader):])
-
-                if (content_length > 0):
-                    content = self.process.stdout.read(content_length).decode(
-                        "UTF-8")
-
-                    payload = None
-                    try:
-                        payload = json.loads(content)
-                        limit = min(len(content), 200)
-                        if payload.get("method") != "window/logMessage":
-                            debug("got json: ", content[0:limit], "...")
-                    except IOError as err:
-                        exception_log("got a non-JSON payload: " + content, err)
-                        continue
-
-                    try:
-                        if "error" in payload:
-                            error = payload['error']
-                            debug("Got error from server: ", error)
-                            sublime.status_message(error.get('message'))
-                        elif "method" in payload:
-                            if "id" in payload:
-                                self.request_handler(payload)
-                            else:
-                                self.notification_handler(payload)
-                        elif "id" in payload:
-                            self.response_handler(payload)
-                        else:
-                            debug("Unknown payload type: ", payload)
-                    except Exception as err:
-                        exception_log("Error handling server payload", err)
-
-            except IOError as err:
-                sublime.status_message("Failure reading LSP server response, exiting")
-                exception_log("Failure reading stdout", err)
-                self.process.terminate()
-                self.process = None
-                return
-
-        debug("LSP stdout process ended.")
-
-    def read_stderr(self):
-        """
-        Reads any errors from the LSP process.
-        """
-        running = True
-        while running:
-            running = self.process.poll() is None
-
-            try:
-                content = self.process.stderr.readline()
-                if not content:
-                    break
-                if settings.log_stderr:
-                    server_log("(stderr): ", content.strip())
-            except IOError as err:
-                exception_log("Failure reading stderr", err)
-                return
-
-        debug("LSP stderr process ended.")
-
-    def response_handler(self, response):
-        handler_id = int(response.get("id"))  # dotty sends strings back :(
-        result = response.get('result', None)
-        if (self.handlers[handler_id]):
-            self.handlers[handler_id](result)
-        else:
-            debug("No handler found for id" + response.get("id"))
-
-    def request_handler(self, request):
-        method = request.get("method")
-        if method == "workspace/applyEdit":
-            apply_workspace_edit(sublime.active_window(),
-                                 request.get("params"))
-        else:
-            debug("Unhandled request", method)
-
-    def notification_handler(self, response):
-        method = response.get("method")
-        if method == "textDocument/publishDiagnostics":
-            Events.publish("document.diagnostics", response.get("params"))
-        elif method == "window/showMessage":
-            sublime.active_window().message_dialog(
-                response.get("params").get("message"))
-        elif method == "window/logMessage" and settings.log_server:
-            server_log(self.process.args[0],
-                       response.get("params").get("message"))
-        else:
-            debug("Unhandled notification:", method)
 
 
 def get_project_path(window: sublime.Window) -> 'Optional[str]':
@@ -809,6 +640,22 @@ def handle_initialize_result(result, client, window, config):
     global didopen_after_initialize
     capabilities = result.get("capabilities")
     client.set_capabilities(capabilities)
+
+    # handle server requests and notifications
+    client.on_request(
+        "workspace/applyEdit",
+        lambda params: apply_workspace_edit(sublime.active_window(), params))
+
+    client.on_notification(
+        "textDocument/publishDiagnostics",
+        lambda params: Events.publish("document.diagnostics", params))
+    client.on_notification(
+        "window/showMessage",
+        lambda params: sublime.active_window().message_dialog(params.get("message")))
+    if settings.log_server:
+        client.on_notification(
+            "window/logMessage",
+            lambda params: server_log(params.get("message")))
 
     # TODO: These handlers is already filtered by syntax but does not need to
     # be enabled 2x per client
